[SearchBundle] Add Index Attributes
