+++ src/Enhavo/Bundle/SearchBundle/Index/Metadata/IndexAttributeDriver.php
@@ -10,8 +10,16 @@ class IndexAttributeDriver implements DriverInterface
     public function loadClass($className): array|null|false
     {
         $reflection = new \ReflectionClass($className);
-
         $properties = [];
+        $attributes = $reflection->getAttributes(Index::class);
+        foreach ($attributes as $attribute) {
+            $arguments = $attribute->getArguments();
+            $options = $arguments[1] ?? [];
+            $options['type'] = $arguments[0];
+            $name = $options['name'];
+            unset($options['name']);
+            $properties[$name] = $options;
+        }
         foreach ($reflection->getProperties() as $property) {
             $attributes = $property->getAttributes(Index::class);
             foreach ($attributes as $attribute) {
