+++ src/Enhavo/Bundle/SearchBundle/Resources/config/services/index.yaml
@@ -6,13 +6,13 @@ services:
 
     Enhavo\Bundle\SearchBundle\Index\Type\ModelIndexType:
         calls:
-            - [setIndexDataProvider, ['@Enhavo\Bundle\SearchBundle\Index\IndexDataProvider']]
+            - [ setIndexDataProvider, [ '@Enhavo\Bundle\SearchBundle\Index\IndexDataProvider' ] ]
         tags:
             - { name: enhavo_search.index }
 
     Enhavo\Bundle\SearchBundle\Index\Type\CollectionIndexType:
         calls:
-            - [setIndexDataProvider, ['@Enhavo\Bundle\SearchBundle\Index\IndexDataProvider']]
+            - [ setIndexDataProvider, [ '@Enhavo\Bundle\SearchBundle\Index\IndexDataProvider' ] ]
         tags:
             - { name: enhavo_search.index }
 
@@ -23,29 +23,3 @@ services:
     Enhavo\Bundle\SearchBundle\Index\Type\TextIndexType:
         tags:
             - { name: enhavo_search.index }
-
-    Enhavo\Component\Metadata\MetadataRepository[Index]:
-        class: Enhavo\Component\Metadata\MetadataRepository
-        arguments:
-            - '@Enhavo\Component\Metadata\MetadataFactory[Index]'
-            - false
-            - false
-
-    Enhavo\Component\Metadata\MetadataFactory[Index]:
-        class: Enhavo\Component\Metadata\MetadataFactory
-        arguments:
-            - 'Enhavo\Bundle\SearchBundle\Index\Metadata\Metadata'
-        calls:
-            - [addDriver, ['@Enhavo\Component\Metadata\Driver\ConfigurationDriver[Index]']]
-            - [addDriver, ['@Enhavo\Bundle\SearchBundle\Index\Metadata\IndexAttributeDriver']]
-            - [addProvider, ['@Enhavo\Bundle\SearchBundle\Index\Metadata\IndexProvider']]
-
-    Enhavo\Component\Metadata\Driver\ConfigurationDriver[Index]:
-        class: Enhavo\Component\Metadata\Driver\ConfigurationDriver
-        arguments:
-            - '%enhavo_search.metadata%'
-
-    Enhavo\Bundle\SearchBundle\Index\Metadata\IndexAttributeDriver:
-
-
-    Enhavo\Bundle\SearchBundle\Index\Metadata\IndexProvider:
